+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to
 - ♿(frontend) focus skip link on headings and skip grid dropzone #1983
 - ♿️(frontend) fix search modal accessibility issues #2054
 - ♿️(frontend) add sr-only format to export download button #2088
+- ♿️(frontend) announce formatting shortcuts for screen readers #2070
 
 ## [v4.8.2] - 2026-03-19